+++ packages/plugins/eslint-plugin/src/index.ts
@@ -87,6 +87,74 @@ const rulePreset = {
   "react-hooks/prefer-use-state-lazy-initialization": "warn",
 } as const satisfies RulePreset;
 
+const jsxPreset = {
+  "jsx/max-depth": "warn",
+  "jsx/no-array-index-key": "warn",
+  "jsx/no-comment-textnodes": "warn",
+  "jsx/no-complicated-conditional-rendering": "warn",
+  "jsx/no-duplicate-key": "error",
+  "jsx/no-leaked-conditional-rendering": "warn",
+  "jsx/no-missing-key": "error",
+  "jsx/no-spreading-key": "warn",
+  "jsx/no-useless-fragment": "warn",
+  "jsx/prefer-shorthand-boolean": "warn",
+  "jsx/prefer-shorthand-fragment": "warn",
+} as const satisfies RulePreset;
+
+const corePreset = {
+  "react/ensure-forward-ref-using-ref": "warn",
+  "react/no-access-state-in-setstate": "error",
+  "react/no-children-count": "warn",
+  "react/no-children-for-each": "warn",
+  "react/no-children-in-void-dom-elements": "warn",
+  "react/no-children-map": "warn",
+  "react/no-children-only": "warn",
+  "react/no-children-prop": "warn",
+  "react/no-children-to-array": "warn",
+  "react/no-class-component": "warn",
+  "react/no-clone-element": "warn",
+  "react/no-component-will-mount": "error",
+  "react/no-component-will-receive-props": "error",
+  "react/no-component-will-update": "error",
+  "react/no-constructed-context-value": "error",
+  "react/no-create-ref": "error",
+  "react/no-direct-mutation-state": "error",
+  "react/no-missing-component-display-name": "warn",
+  "react/no-namespace": "error",
+  "react/no-redundant-should-component-update": "error",
+  "react/no-render-return-value": "error",
+  "react/no-set-state-in-component-did-mount": "warn",
+  "react/no-set-state-in-component-did-update": "warn",
+  "react/no-set-state-in-component-will-update": "warn",
+  "react/no-string-refs": "error",
+  "react/no-unsafe-component-will-mount": "warn",
+  "react/no-unsafe-component-will-receive-props": "warn",
+  "react/no-unsafe-component-will-update": "warn",
+  "react/no-unstable-default-props": "error",
+  "react/no-unstable-nested-components": "warn",
+  "react/no-unused-class-component-members": "warn",
+  "react/no-unused-state": "warn",
+  "react/prefer-destructuring-assignment": "warn",
+} as const satisfies RulePreset;
+
+const domPreset = {
+  "react-dom/no-dangerously-set-innerhtml": "warn",
+  "react-dom/no-dangerously-set-innerhtml-with-children": "error",
+  "react-dom/no-find-dom-node": "error",
+  "react-dom/no-missing-button-type": "warn",
+  "react-dom/no-missing-iframe-sandbox": "warn",
+  "react-dom/no-script-url": "warn",
+  "react-dom/no-unsafe-iframe-sandbox": "warn",
+  "react-dom/no-unsafe-target-blank": "warn",
+} as const satisfies RulePreset;
+
+const hooksPreset = {
+  "react-hooks/ensure-custom-hooks-using-other-hooks": "warn",
+  "react-hooks/ensure-use-callback-has-non-empty-deps": "warn",
+  "react-hooks/ensure-use-memo-has-non-empty-deps": "warn",
+  "react-hooks/prefer-use-state-lazy-initialization": "warn",
+} as const satisfies RulePreset;
+
 const recommendedPreset = {
   // "jsx/max-depth": "warn", // Not enabled because it's style related.
   "jsx/no-array-index-key": "warn",
@@ -209,8 +277,16 @@ export default {
   configs: {
     all: createFlatConfig(allPreset),
     "all-legacy": createLegacyConfig(allPreset),
+    core: createFlatConfig(corePreset),
+    "core-legacy": createLegacyConfig(corePreset),
     debug: createFlatConfig(debugPreset),
     "debug-legacy": createLegacyConfig(debugPreset),
+    dom: createFlatConfig(domPreset),
+    "dom-legacy": createLegacyConfig(domPreset),
+    hooks: createFlatConfig(hooksPreset),
+    "hooks-legacy": createLegacyConfig(hooksPreset),
+    jsx: createFlatConfig(jsxPreset),
+    "jsx-legacy": createLegacyConfig(jsxPreset),
     off: createFlatConfig(offPreset),
     "off-legacy": createLegacyConfig(offPreset),
     recommended: createFlatConfig(recommendedPreset),